Fixing bug with empty sqlfilter
Occurs when there are no create/replace statements in the sql file being executed (oracle_lib.find_entities returns an empty dict).  The second parameter being sent to RunSQL.sql is empty and the USER_ERRORS select statement fails with an error, even though the sql file was executed successfully (which can be confusing to the user).  This fix simple sends an empty string as the second parameter in the case when no entities are returned by oracle_lib.